Have pokemon to integer list working, and retrieving it from list working.
Didnt have time to figure out how to update an array of objects, favorite functionality is not working currently due to the list storing intgers instead of pokemon. Perhaps I can make a secondary list containing the pokemon that have been favorited and then saving that? or turning the list of integers into a map of intgers and booleans. At that point its probably I might as well do a full object, which Im not too sure how to do yet
