+++ app/src/main/java/com/example/team2finalprojectpokedex/Trainer.java
@@ -1,6 +1,15 @@
 package com.example.team2finalprojectpokedex;
 
 
+import android.util.Log;
+
+import androidx.annotation.NonNull;
+
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.IgnoreExtraProperties;
 import com.google.firebase.firestore.ServerTimestamp;
 
@@ -116,5 +125,28 @@ public class Trainer {
                 ", pokedex=" + pokedex +
                 ", timestamp=" + timestamp +
                 '}';
+    }
+    // Should hopefully update trainer each time a new pokemon is made, keeping track of whether they are in their integer array
+    public void updateTrainer(FirebaseUser currentUser, List<Pokemon> pokemons) {
+        //TODO: FIX LOGIC OF MAKING NEW LIST
+        List<Integer> updatedList = new ArrayList<>();
+        Log.d("UPDATE", pokemons.toString());
+        for(Pokemon pokemon: pokemons){
+                updatedList.add(Integer.valueOf(pokemon.getId()));
+        }
+        setPokedex(updatedList);
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        DocumentReference docRef = db.collection("users").document(currentUser.getUid());
+        docRef.update("pokedex", getPokedex()).addOnSuccessListener(new OnSuccessListener<Void>() {
+            @Override
+            public void onSuccess(Void aVoid) {
+                Log.d("UPDATE", "PokeDex Updated in fireStore");
+            }
+        }).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull Exception e) {
+                Log.d("UPDATE", "Failed to update pokedex");
+            }
+        });
     }
 }
